Creation of "Getting_ECOSIS_data_from_API_and_storing_metadata.R" and "Getting_traits_data.R"
Added the R_codes folder which contain the "Getting_ECOSIS_data_from_API_and_storing_metadata.R" and "Getting_traits_data.R". The first code is an API function which reads data from ECOSIS website and collects all data based on leaf (canopy). The code then stores the metadata and spectra files separately. The code can be run periodically to check for any new dataset which will be downloaded to the directory.

The second code takes the data from the directory and gives the names of the datasets containing a particular trait. This is helpful when we want to analyze only a particular trait. Second code still in progress