Updated ECOSIS_implementation_for_Bayesian_ML.R and added an R code to plot the one-to-one plots for predictions/observations for various classifications
The ECOSIS_implementation_for_Bayesian_ML.R now outputs a data frame which has the predictions as well as all the metadata for a particular trait. This code is called inside the R code "Plotting_one_to_one_plots_for_various_classifications.R" which then plots the resulting predictions with the observations for various metadata classifications.